# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: Habitree Backend API
  version: 1.1.0
  description: Aktuelle Definition der REST-Endpunkte des Habitree-Backends (Express + Prisma + JWT).
servers:
  - url: http://iseproject01.informatik.htw-dresden.de:8000
    description: Externer Server
  - url: http://localhost:8000
    description: Lokale Entwicklung
paths:
  /auth/register:
    post:
      summary: Nutzer registrieren
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Registrierung erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400': { description: Ungültige Eingabe }
        '409': { description: E-Mail bereits registriert }
  /auth/login:
    post:
      summary: Nutzer anmelden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401': { description: Ungültige Anmeldedaten }
  /auth/password:
    put:
      summary: Passwort ändern (authentifiziert)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200': { description: Passwort geändert }
        '400': { description: Ungültige Eingabe }
        '401': { description: Nicht autorisiert }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        username:
          type: string
          description: Optional, ansonsten wird die E-Mail-Präfix verwendet.
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword: { type: string }
        newPassword:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      properties:
        message: { type: string }
        token: { type: string }
        user:
          $ref: '#/components/schemas/AuthUser'
    AuthUser:
      type: object
      properties:
        userId: { type: integer }
        email: { type: string, format: email }
        username: { type: string }
...